+++ CHANGELOG.md
@@ -8,6 +8,7 @@
 - Fixed a styling issue with Stripe’s element’s meta card footer.
 - Added `stripe/data/reset` CLI command. ([#42](https://github.com/craftcms/stripe/issues/42))
 - Added support for `--with-fields` option for the `resave/*` commands. ([#55](https://github.com/craftcms/stripe/pull/55))
+- Added support for defining card attributes for the Product, Price and Subscription field layouts. ([#56](https://github.com/craftcms/stripe/pull/56))
 
 ## 1.2.0.2 - 2024-11-08